Migrate pages search functionality into search-bar component
