+++ src/components/page-list.jsx
@@ -1,6 +1,7 @@
 import {dateFormatter, formatSites} from '../scripts/formatters';
 import Loading from './loading';
 import React from 'react';
+import SearchBar from './search-bar/search-bar';
 
 /**
  * These props also inherit from React Router's RouteComponent props
@@ -17,12 +18,6 @@ import React from 'react';
  * @param {PageListProps} props
  */
 export default class PageList extends React.Component {
-  constructor (props) {
-    super(props);
-    this._didSearch = this._didSearch.bind(this);
-    this._dispatchSearch = debounce(this._dispatchSearch.bind(this), 500);
-  }
-
   render () {
     if (!this.props.pages) {
       return <Loading />;
@@ -42,13 +37,7 @@ export default class PageList extends React.Component {
 
     return (
       <div className="container-fluid container-list-view">
-        <div className="row search-bar">
-          <input
-            type="text"
-            placeholder="Search for a URL..."
-            onChange={this._didSearch}
-          />
-        </div>
+        <SearchBar onSearch={this.props.onSearch} />
         {results}
       </div>
     );
@@ -113,28 +102,6 @@ export default class PageList extends React.Component {
 
     this.props.history.push(`/page/${page.uuid}`);
   }
-
-  _didSearch (event) {
-    this._dispatchSearch(event.target.value);
-  }
-
-  _dispatchSearch (url) {
-    if (url) {
-      // If doesn't start with a protocol (or looks like it's going that way),
-      // prefix with an asterisk.
-      if (!/^(\*|\/\/|(h|ht|htt|https?|https?\/|https?\/\/))/.test(url)) {
-        url = url = `*//${url}`;
-      }
-      // If the search is for a domain + TLD, return all paths under it
-      if (/^[\w:*]+(\/\/)?[^/]+$/.test(url)) {
-        url = `${url}*`;
-      }
-    }
-    if (this.props.onSearch) {
-      const query = url ? {url} : null;
-      this.props.onSearch(query);
-    }
-  }
 }
 
 function isInAnchor (node) {
@@ -145,12 +112,4 @@ function isInAnchor (node) {
     return true;
   }
   return isInAnchor(node.parentNode);
-}
-
-function debounce (func, delay) {
-  let timer = null;
-  return function (...args) {
-    clearTimeout(timer);
-    timer = setTimeout(() => func(...args), delay);
-  };
 }
