+++ src/components/search-bar/search-bar.css
@@ -1,0 +1,11 @@
+.search-bar {
+  border-bottom: 1px solid #ccc;
+  margin-bottom: 0.5em;
+}
+
+.search-bar-input {
+  border: none;
+  font-size: 1.25em;
+  padding: 0.5em 15px;
+  width: 100%;
+}
