+++ src/css/styles.css
@@ -160,18 +160,6 @@ body {
   width: 40%;
 }
 
-.search-bar {
-  border-bottom: 1px solid #ccc;
-  margin-bottom: 0.5em;
-}
-
-.search-bar input {
-  border: none;
-  font-size: 1.25em;
-  padding: 0.5em 15px;
-  width: 100%;
-}
-
 /* ===== Page Detail View ===== */
 .page-title {
   font-size: 1.5em;
